--- src/main/java/nextstep/laddergame/domain/Goal.java
@@ -1,5 +1,0 @@
-package nextstep.laddergame.domain;
-
-public record Goal(String goal) {
-
-}
