+++ src/main/java/nextstep/laddergame/domain/ladder/Goal.java
@@ -1,0 +1,5 @@
+package nextstep.laddergame.domain.ladder;
+
+public record Goal(String goal) {
+
+}
