--- src/test/java/nextstep/laddergame/domain/ParticipantTest.java
@@ -1,5 +1,0 @@
-package nextstep.laddergame.domain;
-
-class ParticipantTest {
-
-}
