+++ src/test/java/nextstep/laddergame/domain/ladder/LadderTest.java
@@ -1,4 +1,4 @@
-package nextstep.laddergame.domain;
+package nextstep.laddergame.domain.ladder;
 
 import static nextstep.laddergame.builder.LadderBuilder.aLadderBuilder;
 import static nextstep.laddergame.builder.LadderBuilder.aLadderResult;
