+++ src/test/java/nextstep/laddergame/domain/participant/ParticipantTest.java
@@ -1,0 +1,5 @@
+package nextstep.laddergame.domain.participant;
+
+class ParticipantTest {
+
+}
