+++ src/test/java/nextstep/laddergame/domain/participant/ParticipantsTest.java
@@ -1,7 +1,6 @@
-package nextstep.laddergame.domain;
+package nextstep.laddergame.domain.participant;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import nextstep.laddergame.domain.participant.Participants;
 import org.junit.jupiter.api.Test;
 
 class ParticipantsTest {
